crop_prediction_api/requirements.txt
Flask>=3.0,<4
Flask-Cors>=4.0,<5
Flask-Limiter>=3.8,<4
joblib>=1.4,<2
numpy>=1.26,<3
Pillow>=10.0,<13
pydantic>=2.7,<3
python-dotenv>=1.0,<2
scikit-learn==1.2.2
torch>=2.2,<3
tensorflow>=2.15,<3
tf-keras>=2.15,<3
gunicorn>=22.0,<23
